+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## Unreleased
 
+**Breaking Changes**:
+
+- Stop producing generic metrics (spans/transactions metric buckets) to Kafka and remove the
+  `metrics_generic` / `ingest-performance-metrics` topic. Session (release health) metrics are
+  unchanged. ([#6388](https://github.com/getsentry/relay/pull/6388))
+
 **Features**:
 
 - Extend cookie scrubbing to response cookies. ([#6391](https://github.com/getsentry/relay/pull/6391))